Module codedpedia::render

This module exports items related to rendering components.

Re-exports

pub use html::Html;

pub use markdown::Markdown;

pub use text::Text;

Modules

html

This module provides utilities about the HTML rendering format.

markdown

This module provides utilities about the markdown rendering format.

text

This module provides utilities about the plaintext rendering format.

Structs

Context

Context data of a session of rendering.

RenderAsDisplay

A helper type that will render a component using Rust’s fmt::Display trait.

Renderer

A proxy formatter for rendering using Rust’s formatting and a render format.

Traits

Format

A rendering format: e.g. HTML, Markdown, etc.

FullRender

An auto-trait for components that implement rendering for all supported formats.

Render

General rendering trait. A component implements this trait when it can renders using the given format W.

Scope

A scope of a render format. The scope allows a component to change rendering configuration by entering the scope.

Type Definitions

DynFullComponent

A dynamic trait object for thread-safe components with full render support.

DynFullComponentUnsync

A dynamic trait object for thread-unsafe components with full render support.